Extract exceptions in UniquePersonList
Model API refers to UniquePersonList.DuplicatePersonException and
UniquePersonList.PersonNotFoundException.

UniquePersonList is a class inside the model component which need not
be exposed to classes outside the Model component.

Let's extract these two exception as standalone classes so that the
Model API no longer exposes the UniquePersonList class.
