+++ src/main/java/seedu/address/logic/commands/DeleteCommand.java
@@ -5,7 +5,7 @@ import seedu.address.commons.core.UnmodifiableObservableList;
 import seedu.address.commons.core.index.Index;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.person.ReadOnlyPerson;
-import seedu.address.model.person.UniquePersonList.PersonNotFoundException;
+import seedu.address.model.person.exceptions.PersonNotFoundException;
 
 /**
  * Deletes a person identified using it's last displayed index from the address book.
