+++ src/main/java/seedu/address/logic/commands/EditCommand.java
@@ -20,7 +20,8 @@ import seedu.address.model.person.Name;
 import seedu.address.model.person.Person;
 import seedu.address.model.person.Phone;
 import seedu.address.model.person.ReadOnlyPerson;
-import seedu.address.model.person.UniquePersonList;
+import seedu.address.model.person.exceptions.DuplicatePersonException;
+import seedu.address.model.person.exceptions.PersonNotFoundException;
 import seedu.address.model.tag.Tag;
 
 /**
@@ -74,9 +75,9 @@ public class EditCommand extends Command {
 
         try {
             model.updatePerson(personToEdit, editedPerson);
-        } catch (UniquePersonList.DuplicatePersonException dpe) {
+        } catch (DuplicatePersonException dpe) {
             throw new CommandException(MESSAGE_DUPLICATE_PERSON);
-        } catch (UniquePersonList.PersonNotFoundException pnfe) {
+        } catch (PersonNotFoundException pnfe) {
             throw new AssertionError("The target person cannot be missing");
         }
         model.updateFilteredListToShowAll();
