+++ src/main/java/seedu/address/model/Model.java
@@ -4,9 +4,8 @@ import java.util.Set;
 
 import seedu.address.commons.core.UnmodifiableObservableList;
 import seedu.address.model.person.ReadOnlyPerson;
-import seedu.address.model.person.UniquePersonList;
-import seedu.address.model.person.UniquePersonList.DuplicatePersonException;
-import seedu.address.model.person.UniquePersonList.PersonNotFoundException;
+import seedu.address.model.person.exceptions.DuplicatePersonException;
+import seedu.address.model.person.exceptions.PersonNotFoundException;
 
 /**
  * The API of the Model component.
@@ -19,10 +18,10 @@ public interface Model {
     ReadOnlyAddressBook getAddressBook();
 
     /** Deletes the given person. */
-    void deletePerson(ReadOnlyPerson target) throws UniquePersonList.PersonNotFoundException;
+    void deletePerson(ReadOnlyPerson target) throws PersonNotFoundException;
 
     /** Adds the given person */
-    void addPerson(ReadOnlyPerson person) throws UniquePersonList.DuplicatePersonException;
+    void addPerson(ReadOnlyPerson person) throws DuplicatePersonException;
 
     /**
      * Replaces the given person {@code target} with {@code editedPerson}.
@@ -32,7 +31,7 @@ public interface Model {
      * @throws PersonNotFoundException if {@code target} could not be found in the list.
      */
     void updatePerson(ReadOnlyPerson target, ReadOnlyPerson editedPerson)
-            throws UniquePersonList.DuplicatePersonException, UniquePersonList.PersonNotFoundException;
+            throws DuplicatePersonException, PersonNotFoundException;
 
     /** Returns the filtered person list as an {@code UnmodifiableObservableList<ReadOnlyPerson>} */
     UnmodifiableObservableList<ReadOnlyPerson> getFilteredPersonList();
