+++ src/main/java/seedu/address/model/person/exceptions/PersonNotFoundException.java
@@ -1,0 +1,6 @@
+package seedu.address.model.person.exceptions;
+
+/**
+ * Signals that the operation is unable to find the specified person.
+ */
+public class PersonNotFoundException extends Exception {}
